program update
The program is fully tested and work as intended. Now the program will synchronize the benchmark and target database for the given table names provided. Note that this program is designed for Oracle database (SQL syntax might be different for other databases) and the syntax might be slightly different between each company's naming convention. In the next couple of days, I will try to switch to Maven instead of Makefile.